RECURSIVE NIGHTMARE OH MY GOD
