+++ contracts/__init__.py
@@ -1,14 +1,18 @@
-import time
 from .classes import SeasonDB
 from .seasons import Winter2025
-from .master import MasterDB
 from .classes import *  # noqa: F403
-from utils import config
+from common import config
+import time
+import discord
+import utils
 
 AVAILABLE_SEASONS = ["Winter 2025"]
 
 
-async def get_season_db(season: str = config.active_season) -> SeasonDB:
+async def get_season_db(season: str = None) -> SeasonDB:
+	if season is None:
+		season = config.active_season
+
 	if season not in AVAILABLE_SEASONS:
 		raise ValueError(f"Invalid season: {season}")
 
@@ -17,10 +21,10 @@ async def get_season_db(season: str = config.active_season) -> SeasonDB:
 			return await Winter2025.get_database()
 
 
-master_db: MasterDB = None
+async def sync_season_db(season: str = None) -> float:  # Returns duration of sync
+	if season is None:
+		season = config.active_season
 
-
-async def sync_season_db(season: str = config.active_season) -> float:  # Returns duration of sync
 	if season not in AVAILABLE_SEASONS:
 		raise ValueError(f"Invalid season: {season}")
 
@@ -32,3 +36,13 @@ async def sync_season_db(season: str = config.active_season) -> float:  # Return
 			await Winter2025.sync_to_latest(db)
 
 	return time.perf_counter() - start
+
+
+async def usernames_autocomplete(ctx: discord.AutocompleteContext):
+	season_db = await get_season_db()
+	return await utils.contracts.get_usernames(season_db, query=ctx.value.strip(), limit=25)
+
+
+async def reps_autocomplete(ctx: discord.AutocompleteContext):
+	season_db = await get_season_db()
+	return await utils.contracts.get_reps(season_db, query=ctx.value.strip(), limit=25)
